config/requirements.txt
requests
psycopg2
mysql-connector-python
python-dotenv
pydantic-settings
